--- src-final/app/layout.tsx
@@ -1,28 +1,0 @@
-import type { Metadata } from "next";
-import { Inter } from "next/font/google";
-import "./globals.css";
-import { Toaster } from "react-hot-toast";
-
-const inter = Inter({ subsets: ["latin"] });
-
-export const metadata: Metadata = {
-  title: "Supabase User",
-};
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <html lang="en">
-      <body className={inter.className}>
-        <main className="flex flex-col min-h-screen bg-zinc-800 text-zinc-200 justify-center items-center pb-40">
-          {children}
-        </main>
-
-        <Toaster />
-      </body>
-    </html>
-  );
-}
