--- src-final/app/protected/page.tsx
@@ -1,5 +1,0 @@
-function ProtectedPage() {
-  return <div>This page is protected</div>;
-}
-
-export default ProtectedPage;
