+++ src/app/forgot-password/page.tsx
@@ -1,23 +1,23 @@
 "use client";
 
-import { loginAction } from "@/actions/users";
+import { forgotPasswordAction } from "@/actions/users";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
 import { useTransition } from "react";
 import toast from "react-hot-toast";
 
-function LoginPage() {
+function ForgotPassword() {
   const router = useRouter();
 
   const [isPending, startTransition] = useTransition();
 
-  const handleClickLoginButton = (formData: FormData) => {
+  const handleClickSendEmail = (formData: FormData) => {
     startTransition(async () => {
-      const { errorMessage } = await loginAction(formData);
+      const { errorMessage } = await forgotPasswordAction(formData);
       if (errorMessage) {
         toast.error(errorMessage);
       } else {
-        toast.success("Logged in");
+        toast.success("Check your email to reset your password.");
         router.push("/");
       }
     });
@@ -27,9 +27,9 @@ function LoginPage() {
     <div>
       <form
         className="flex flex-col bg-emerald-700 w-96 p-8 rounded-lg gap-y-4"
-        action={handleClickLoginButton}
+        action={handleClickSendEmail}
       >
-        <h1 className="text-4xl text-center mb-8">Login</h1>
+        <h1 className="text-4xl text-center mb-8">Forgot Password</h1>
 
         <input
           type="text"
@@ -38,25 +38,18 @@ function LoginPage() {
           placeholder="Email"
           disabled={isPending}
         />
-        <input
-          type="password"
-          name="password"
-          className="rounded-lg p-2 text-black"
-          placeholder="Password"
-          disabled={isPending}
-        />
 
         <button
           className="bg-black rounded-lg p-2 mt-4 mb-2"
           disabled={isPending}
         >
-          {isPending ? "Logging in..." : "Login"}
+          {isPending ? "Sending..." : "Send Email"}
         </button>
 
         <p className="text-center">
-          Don't have an account?{" "}
-          <Link href="/create-account" className="underline">
-            Create Account
+          Already have an account?{" "}
+          <Link href="/login" className="underline">
+            Login
           </Link>
         </p>
       </form>
@@ -64,4 +57,4 @@ function LoginPage() {
   );
 }
 
-export default LoginPage;
+export default ForgotPassword;
