+++ src/app/login/page.tsx
@@ -1,9 +1,27 @@
 "use client";
 
+import { loginAction } from "@/actions/users";
 import Link from "next/link";
+import { useRouter } from "next/navigation";
+import { useTransition } from "react";
+import toast from "react-hot-toast";
 
 function LoginPage() {
-  const handleClickLoginButton = (formData: FormData) => {};
+  const router = useRouter();
+
+  const [isPending, startTransition] = useTransition();
+
+  const handleClickLoginButton = (formData: FormData) => {
+    startTransition(async () => {
+      const { errorMessage } = await loginAction(formData);
+      if (errorMessage) {
+        toast.error(errorMessage);
+      } else {
+        toast.success("Logged in");
+        router.push("/");
+      }
+    });
+  };
 
   return (
     <div>
@@ -18,20 +36,33 @@ function LoginPage() {
           name="email"
           className="rounded-lg p-2 text-black"
           placeholder="Email"
+          disabled={isPending}
         />
         <input
           type="password"
           name="password"
           className="rounded-lg p-2 text-black"
           placeholder="Password"
+          disabled={isPending}
         />
 
-        <button className="bg-black rounded-lg p-2 mt-4 mb-2">Login</button>
+        <button
+          className="bg-black rounded-lg p-2 mt-4 mb-2"
+          disabled={isPending}
+        >
+          {isPending ? "Logging in..." : "Login"}
+        </button>
 
         <p className="text-center">
           Don't have an account?{" "}
           <Link href="/create-account" className="underline">
             Create Account
+          </Link>
+        </p>
+        <p className="text-center">
+          Forgot your password?{" "}
+          <Link href="/forgot-password" className="underline">
+            Click here
           </Link>
         </p>
       </form>
